add CRUD user and CRUD role_tpu

+++ api_emakam/app/Http/Controllers/AdminPusatController.php
@@ -13,11 +13,20 @@ class AdminPusatController extends Controller
 {
     function view(){
     	$user = DB::table('user')
-            ->join('role_tpu', 'user.id_user', '=', 'role_tpu.id_user')
-            ->select('user.*', 'role_tpu.*')
-            ->get();
+        ->join('role_tpu', 'user.id_user', '=', 'role_tpu.id_user')
+        ->select('user.*', 'role_tpu.*')
+        ->get();
 
-		return response()->json($user);
+        return response()->json($user);
+    }
+
+    function constraint_user(){
+        $user = DB::table('user')
+        ->join('role_tpu', 'user.id_user', '=', 'role_tpu.id_user')
+        ->join('tpu', 'role_tpu.id_tpu', '=', 'tpu.id_tpu')
+        ->select('user.*', 'role_tpu.*','tpu.*')
+        ->get();
+        return response()->json($user);
     }
 
     function view_tpu(){
@@ -28,10 +37,10 @@ class AdminPusatController extends Controller
     
     function view_penghunimakam(){
         $view = DB::table('penghuni_makam')
-            ->join('makam', 'penghuni_makam.id_makam', '=', 'makam.id_makam')
-            ->select('penghuni_makam.*', 'makam.*')
-            ->get();
-		return response()->json($view);
+        ->join('makam', 'penghuni_makam.id_makam', '=', 'makam.id_makam')
+        ->select('penghuni_makam.*', 'makam.*')
+        ->get();
+        return response()->json($view);
     }
 
     function create_penghunimakam(Request $request){
@@ -93,16 +102,16 @@ class AdminPusatController extends Controller
 
     function view_blok(){
     	$view = DB::table('blok_makam')
-            ->join('tpu', 'blok_makam.id_tpu', '=', 'tpu.id_tpu')
-            ->select('blok_makam.*', 'tpu.*')
-            ->get();
-		return response()->json($view);
+        ->join('tpu', 'blok_makam.id_tpu', '=', 'tpu.id_tpu')
+        ->select('blok_makam.*', 'tpu.*')
+        ->get();
+        return response()->json($view);
     }
 
 
     function view_role_tpu(){
-    	$role_tpu = Role_tpu::all();
-		return response()->json($role_tpu);
+        $role_tpu = Role_tpu::all();
+        return response()->json($role_tpu);
     }
 
     function create_user(Request $request)
@@ -110,8 +119,8 @@ class AdminPusatController extends Controller
         $username = $request->input('username');
         $password = $request->input('password');
         $role = $request->input('role');
-        $id_tpu = $request->input('id_tpu');
-        $id_user = "";
+        //$id_tpu = $request->input('id_tpu');
+        //$id_user = "";
 
         $user = new User([
             'username' => $username,
@@ -119,27 +128,81 @@ class AdminPusatController extends Controller
             'role' => $role
         ]);
 
-        $role_tpu = new Role_tpu([
-            'id_tpu' => $id_tpu,
-            'id_user' => $id_user
-        ]);
+        $user->save();
     }
 
-    function update_user(Request $request, $id)
+    function update_user(Request $request, $id_user)
     {
-        $user = User::findOrFail($id);
-        $user->update($request->all());
+        $user = User::findOrFail($id_user);
 
-        return $user;
+
+        $username = $request->input('username');
+        $password = $request->input('password');
+        $role = $request->input('role');
+
+        $user->update([
+            'username' => $request->input('username'),
+            'password' => $request->input('password'),
+            'role' => $request->input('role')
+        ]);
+        $role_tpu->save();
+
+        return response()->json($username);
     }
 
     function delete_user(Request $request, $id_user)
     {
-    	$role_tpu = Role_tpu::findOrFail($id_user);
-        $user = User::findOrFail($id_user);
+        $user = DB::table('user')->select('id_user')->where('id_user', '=', $id_user)->value('id_user');
+        $role_tpu = DB::table('role_tpu')->select('id_user')->where('id_user', '=', $id_user)->value('id_user');
+        if(empty($user)){
+            $msg = ([
+                'msg' => "Data User tidak ditemukan !"
+            ]);
+        } else {
+            if(empty($role_tpu)){
+                $user = User::findOrFail($id_user);
+                $user->delete();
+                $msg = ([
+                    'msg' => "User berhasil dihapus"
+                ]);
+            } else {
+                $msg = ([
+                    'msg' => "Anda harus menghapus hak akses user terlebih dahulu !"
+                ]);
+            }
+        }
+        return response()->json($msg);
+    }
+
+    function create_role_tpu(Request $request)
+    {
+        $id_tpu = $request->input('id_tpu');
+        $id_user = $request->input('id_user');
+
+        $role_tpu = new Role_tpu([
+            'id_tpu' => $id_tpu,
+            'id_user' => $id_user
+        ]);
+
+        $role_tpu->save();
+
+        return response()->json($role_tpu);
+    }
+
+    function update_role_tpu(Request $request, $id_user)
+    {
+        $role_tpu = Role_tpu::findOrFail($id_user);
+        
+        $role_tpu->update($request->all());
+
+        return response()->json(['msg' => "Hak Akses Berhasil di update!"]);
+    }
+
+    function delete_role_tpu(Request $request, $id_user)
+    {
+        $role_tpu = Role_tpu::findOrFail($id_user);
         $role_tpu->delete();
-        $user->delete();
-       
-        return 204;
+
+        return response()->json(['msg' => "Hak Akses Berhasil dihapus!"]);
     }
 }
